feat: 整合20大陆WZ数据 + 脚本/代码/数据库更新
- WZ数据重组: wz/迁移至wz-zh-CN/，新增大量自定义装备/道具/怪物XML
- 新增装备定制属性系统(equip_custom_attr)
- 新增匠人街/吸怪VIP等脚本
- 新增拍卖脚本优化、每日活跃、宝石锻造、炼金系统
- 新增破功系统修复、伤害加成配置
- Flyway迁移: V1.11.34~V1.11.37
- 服务端Java: 装备伤害加成/进阶/强化配置管理
- .gitignore: 排除临时文件(*.bak/*.log/__pycache__/.agents/)

+++ gms-server/src/main/java/org/gms/net/server/channel/handlers/TakeDamageHandler.java
@@ -70,7 +70,7 @@ public final class TakeDamageHandler extends AbstractPacketHandler {
         p.readInt();
         byte damagefrom = p.readByte();
         p.readByte(); //Element
-        long damage = p.readLong();
+        long damage = p.readInt(); // 客户端以32位int发送伤害值，服务端用long存储
         int oid = 0, monsteridfrom = 0, pgmr = 0, direction = 0;
         int pos_x = 0, pos_y = 0, fake = 0;
         boolean is_pgmr = false, is_pg = true, is_deadly = false;
